+++ docs/conf.py
@@ -19,13 +19,13 @@ sys.path.insert(0, os.path.abspath('../'))
 
 project = 'EBCIC'
 copyright = (
-    '2020-2021 '
+    '2020-2022 '
     'National Institute of Advanced Industrial Science and Technology (AIST)'
 )
 author = 'Kazukuni KOBARA, AIST'
 
 # The full version, including alpha/beta/rc tags
-release = '0.0.2'
+release = '0.0.3'
 
 
 # -- General configuration ---------------------------------------------------
@@ -61,4 +61,4 @@ html_theme = "nature"
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+# html_static_path = ['_static']